fixed clang compile locally

+++ src/Generating/BioGen.cpp
@@ -269,7 +269,7 @@ void cBioGenCheckerboard::GenBiomes(cChunkCoords a_ChunkCoords, cChunkDef::Biome
 		{
 			int Add = cChunkDef::Width * a_ChunkCoords.m_ChunkX + x;
 			size_t BiomeIdx = static_cast<size_t>((((Base + Add / m_BiomeSize) % m_BiomesCount) + m_BiomesCount) % m_BiomesCount);  // Need to add and modulo twice because of negative numbers
-			a_BiomeMap[x + cChunkDef::Width * z] = m_Biomes[BiomeIdx];
+			a_BiomeMap[static_cast<size_t>(x + cChunkDef::Width * z)] = m_Biomes[BiomeIdx];
 		}
 	}
 }
