+++ src/Items/ItemPainting.h
@@ -84,7 +84,7 @@ public:
 			"BurningSkull"
 		};
 
-		auto PaintingTitle = gPaintingTitlesList[a_World->GetTickRandomNumber(gPaintingTitlesList.size() - 1)];
+		auto PaintingTitle = gPaintingTitlesList[static_cast<size_t>(a_World->GetTickRandomNumber(gPaintingTitlesList.size() - 1))];
 		auto Painting = std::make_unique<cPainting>(PaintingTitle, a_ClickedBlockFace, PlacePos);
 		auto PaintingPtr = Painting.get();
 		if (!PaintingPtr->Initialize(std::move(Painting), *a_World))
